requirements.txt
Flask==3.0.3
gunicorn==21.2.0
pandas==2.2.2
praw==7.7.1
python-dotenv==1.0.1
torch==2.0.1
tensorflow==2.13.0
numpy==1.24.3

